safari-534.53-branch/LayoutTests/media/broken-video-expected.txt

Test that QuickTime file with broken content generates an error.

EVENT(error)
EXPECTED (video.error == '[object MediaError]') OK
END OF TEST
